exercise4.21

+++ part4/blog_list/controllers/blogs.js
@@ -39,8 +39,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.decode(request.token, process.env.SECRET)
+  const blogToDelete = await Blog.findById(request.params.id)
+  const blogsOwnerId = blogToDelete.user.toString()
+  //console.log(decodedToken.id, blogToDelete.user.toString(), decodedToken.id === blogToDelete.user.toString());
+
+  if (decodedToken === blogToDelete.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    response.status(401).json({ error: 'unauthorized deletion' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
